fix: top right btn in home
